chore(executor): extract test-utils
