+++ crates/executor/src/lib.rs
@@ -21,3 +21,7 @@ pub mod executor;
 /// ExecutorFactory impl
 pub mod factory;
 pub use factory::Factory;
+
+#[cfg(any(test, feature = "test-utils"))]
+/// Common test helpers for mocking out executor and executor factory
+pub mod test_utils;
